data-processing-statistics/requirements.txt
numpy
simplejson
paho-mqtt
matplotlib
weasyprint
cherrypy
pymongo
stats
pytz